exception changes

+++ src/main/java/com/example/productapplication/controller/productcontroller.java
@@ -1,6 +1,7 @@
 package com.example.productapplication.controller;
 
 import com.example.productapplication.dto.createproductrequestdto;
+import com.example.productapplication.exception.ProductNotFoundException;
 import com.example.productapplication.model.product;
 import com.example.productapplication.service.FakeStoreProductService;
 import org.springframework.http.ResponseEntity;
@@ -17,11 +18,14 @@ public class productcontroller {
     }
 
     @GetMapping("/products/{id}")
-    public  ResponseEntity<product> getproductById(@PathVariable("id") Integer id){
-        if(id == null){
-            throw new IllegalArgumentException("id can not be null");
+    public  ResponseEntity<product> getproductById(@PathVariable("id") Integer id) throws ProductNotFoundException {
+        if(id <= 0){
+            throw new IllegalArgumentException("id is Invalid");
         }
         product product = service.getProductById(id);
+        if(product == null){
+            throw new ProductNotFoundException("Product not Found");
+        }
         return ResponseEntity.ok(product);
     }
 
@@ -37,6 +41,15 @@ public class productcontroller {
 
     @PostMapping("/products")
     public ResponseEntity<product> createproduct(@RequestBody createproductrequestdto request){
+        if (request.getTitle() == null) {
+            throw new IllegalArgumentException("Title cannot be null");
+        }
+        if (request.getDescription() == null){
+            throw new IllegalArgumentException("Description cannot be null");
+        }
+        if(request.getImageURL() == null){
+            throw new IllegalArgumentException("ImageURL cannot be null");
+        }
 
         return ResponseEntity.ok(service.createProduct(request.getTitle(),
                 request.getImageURL(), request.getCategory().getTitle(),request.getDescription()));
@@ -52,9 +65,16 @@ public class productcontroller {
     //---------------------------------------------------------------------------------------------------
 
     @DeleteMapping("/products/{id}")
-    public ResponseEntity<String> deleteProductById(@PathVariable("id") Integer id){
-        service.deleteProductById(id);
-        return ResponseEntity.ok("Product deleted successfully! :)");
+    public ResponseEntity<product> deleteProductById(@PathVariable("id") Integer id) throws  ProductNotFoundException{
+        if(id <= 0){
+            throw new IllegalArgumentException("id is Invalid");
+        }
+
+        product product = service.deleteProductById(id);
+        if(product == null){
+            throw new ProductNotFoundException("Product not Found");
+        }
+        return ResponseEntity.ok(product);
 
     }
